back-test

+++ MicroserviceUser/src/main/java/com/helloIftekhar/springJwt/controller/UserController.java
@@ -56,8 +56,7 @@ public class UserController {
     @DeleteMapping("/user/{ids}")
     public ResponseEntity<String> deleteUsersByIds(@PathVariable List<String> ids) {
         for (String id : ids) {
-            if (userService.existsById(id)) {
-                userService.deleteUserById(id);
+            if (userService.existsById(id)) {userService.deleteUserById(id);
             }
         }
         return new ResponseEntity<>(HttpStatus.OK);
